Add 2024 day 6 part 2
This was not fun. I don't feel like my EXPAND function is very lispy and the
graphy cycle detection ended up being a big time sink because I thought it was
detecting them wrongly -- test input was good, but real input was too low.

The quick, dirty, brute force boiling the ocean strategy I was doing was the
real culprit.