Add 2024 day 12 part 2
This was a doozy for me.

Because the regions that I generated in part one included interior matches --
not just the perimeter -- I decided to count the corners. I needed an
operation that could work on each face / pixel of my region tree. After much
research I decided to try a modified 2x2 marching square that marches around a
pixel before moving to the next one.

It was interesting learning and research but frustrating that I couldn't get
it with much of what I made in part 1.